Fix ?to= date filter dropping the final second's sub-second rows
_parse_date_epoch(end_of_day=True) resolves a bare `to=YYYY-MM-DD` to the
fractional end-of-day epoch (…23:59:59.999999) precisely so `epoch <= bound`
keeps millisecond rows in the closing second. But both DB query builders then
did `int(end_epoch)`, truncating the bound back to …59 and re-dropping
23:59:59.001–.999 — defeating the fix end-to-end for every probe that logs
sub-second timestamps.

This hit the JSON read API (fetch_readings) and every export path (CSV, xlsx
and count share _export_where). Switch the four absolute-bound sites from
int() to float() so the fractional bound survives to the query; float() keeps
the defensive numeric coercion and is exact for whole-second bounds.

Add a regression test covering both the read and export paths (and confirming
the next day stays excluded); it fails on the old int() code.

+++ core/db.py
@@ -490,9 +490,13 @@ class Database:
         if cutoff is not None:
             clauses.append("epoch >= ?"); params_list.append(cutoff)
         if start_epoch is not None:
-            clauses.append("epoch >= ?"); params_list.append(int(start_epoch))
+            clauses.append("epoch >= ?"); params_list.append(float(start_epoch))
         if end_epoch is not None:
-            clauses.append("epoch <= ?"); params_list.append(int(end_epoch))
+            # float(), not int(): a `to=YYYY-MM-DD` filter resolves to the
+            # fractional end-of-day epoch (…23:59:59.999999) so that
+            # `epoch <= bound` keeps the final second's millisecond rows. int()
+            # truncated the bound to …59 and silently dropped 23:59:59.001–.999.
+            clauses.append("epoch <= ?"); params_list.append(float(end_epoch))
         if probe_id:
             clauses.append("probe_id = ?"); params_list.append(probe_id)
         where = ("WHERE " + " AND ".join(clauses)) if clauses else ""
@@ -642,9 +646,13 @@ class Database:
         if cutoff is not None:
             clauses.append("epoch >= ?"); params_list.append(cutoff)
         if start_epoch is not None:
-            clauses.append("epoch >= ?"); params_list.append(int(start_epoch))
+            clauses.append("epoch >= ?"); params_list.append(float(start_epoch))
         if end_epoch is not None:
-            clauses.append("epoch <= ?"); params_list.append(int(end_epoch))
+            # float(), not int(): a `to=YYYY-MM-DD` filter resolves to the
+            # fractional end-of-day epoch (…23:59:59.999999) so that
+            # `epoch <= bound` keeps the final second's millisecond rows. int()
+            # truncated the bound to …59 and silently dropped 23:59:59.001–.999.
+            clauses.append("epoch <= ?"); params_list.append(float(end_epoch))
         if probe_id:
             clauses.append("probe_id = ?"); params_list.append(probe_id)
         where = ("WHERE " + " AND ".join(clauses)) if clauses else ""
